Fixed issue #22
