+++ AxiomGL/src/main/java/com/axiom/engine/IGame.java
@@ -1,11 +1,11 @@
 package com.axiom.engine;
 
-import com.axiom.engine.input.MouseListener;
+import com.axiom.engine.input.InputHandler;
 
 public interface IGame {
     void init(Window window) throws Exception;
     void render(Window window);
     void cleanup();
-	void update(float interval, MouseListener mouseInput);
-	void input(Window window, MouseListener mouseInput);
+	void update(float interval, InputHandler keyboardInput);
+	void input(Window window, InputHandler keyboardInput);
 }
