+++ AxiomGL/src/main/java/com/axiom/engine/input/InputHandler.java
@@ -1,9 +1,15 @@
 package com.axiom.engine.input;
 
 import org.lwjgl.glfw.*;
+
+import com.axiom.engine.Window;
+
 import static org.lwjgl.glfw.GLFW.*;
 
-public final class KeyboardListener {
+import org.joml.Vector2d;
+import org.joml.Vector2f;
+
+public final class InputHandler {
 	private static final int KEYBOARD_SIZE = 512;
 	private static final int MOUSE_SIZE = 16;
 
@@ -18,16 +24,26 @@ public final class KeyboardListener {
 	private long lastMouseNS = 0;
 	private long mouseDoubleClickPeriodNS = 1000000000 / 5; // 5th of a second
 															// for double click.
-
 	private int NO_STATE = -1;
 
+    private final Vector2d previousPos;
+    private final Vector2d currentPos;
+    private final Vector2f displVec;
+    
+    private boolean inWindow = false;
+
+    public InputHandler() {
+        previousPos = new Vector2d(-1, -1);
+        currentPos = new Vector2d(0, 0);
+        displVec = new Vector2f();
+    }
 	public GLFWScrollCallback scroll = new GLFWScrollCallback()
     {
-    	public void invoke(long window, double xoffset, double yoffset)
-    	{
-    		scrollStates[0] = xoffset;
-    		scrollStates[1] = yoffset;
-    	}
+	    	public void invoke(long window, double xoffset, double yoffset)
+	    	{
+	    		scrollStates[0] = xoffset;
+	    		scrollStates[1] = yoffset;
+	    	}
     };
 
 	public GLFWKeyCallback keyboard = new GLFWKeyCallback()
@@ -37,26 +53,27 @@ public final class KeyboardListener {
             activeKeys[key]=action!=GLFW_RELEASE;keyStates[key]=action;
         }
     };
-    /*
+    
 	public GLFWMouseButtonCallback mouse = new GLFWMouseButtonCallback()
     {
         public void invoke(long window, int button, int action, int mods)
         {
-            activeMouseButtons[button]=action!=GLFW_RELEASE;getMouseButtonStates()[button]=action;}};
-	*/
-	private void resetKeyboard() {
+            activeMouseButtons[button]=action!=GLFW_RELEASE;getMouseButtonStates()[button]=action;}
+    };
+	
+	public void resetKeyboard() {
 		for (int i = 0; i < keyStates.length; i++) {
 			keyStates[i] = NO_STATE;
 		}
 	}
 
-	private void resetScroll(){
+	public void resetScroll(){
 		for( int i = 0; i < scrollStates.length; i ++ ){
 			scrollStates[i] = 0.0;
 		}
 	}
 	
-	private void resetMouse() {
+	public void resetMouse() {
 		for (int i = 0; i < getMouseButtonStates().length; i++) {
 			getMouseButtonStates()[i] = NO_STATE;
 		}
@@ -116,5 +133,41 @@ public final class KeyboardListener {
 	
 	public double[] getScrollStates(){
 		return scrollStates;
+	}
+	
+    public void init(Window window) {
+		glfwSetKeyCallback(window.getWindowHandle(), this.keyboard);
+		glfwSetScrollCallback(window.getWindowHandle(), this.scroll);
+        glfwSetCursorPosCallback(window.getWindowHandle(), (windowHandle, xpos, ypos) -> {
+            currentPos.x = xpos;
+            currentPos.y = ypos;
+        });
+        glfwSetCursorEnterCallback(window.getWindowHandle(), (windowHandle, entered) -> {
+            inWindow = entered;
+        });
+        glfwSetMouseButtonCallback(window.getWindowHandle(), this.mouse);
+    }
+
+	public Vector2f getDisplVec() {
+		return displVec;
+	}
+
+	public void input(Window window) {
+        displVec.x = 0;
+        displVec.y = 0;
+        if (previousPos.x > 0 && previousPos.y > 0 && inWindow) {
+            double deltax = currentPos.x - previousPos.x;
+            double deltay = currentPos.y - previousPos.y;
+            boolean rotateX = deltax != 0;
+            boolean rotateY = deltay != 0;
+            if (rotateX) {
+                displVec.y = (float) deltax;
+            }
+            if (rotateY) {
+                displVec.x = (float) deltay;
+            }
+        }
+        previousPos.x = currentPos.x;
+        previousPos.y = currentPos.y;
 	}
 }
